Fixed minor type-casting warnings. Added using of delay loaded DLLs.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@8811 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/Buffer.h
@@ -130,8 +130,8 @@ private:
 
 // Statics
 public:
-	static const DWORD	MAX_RECV_SIZE	= 1024ul * 16ul;	// Recieve up to 16KB blocks from the socket
-	static const UINT	ZLIB_CHUNK_SIZE	= 1024u;			// Chunk size for ZLib compression/decompression
+	static const size_t	MAX_RECV_SIZE	= 1024ul * 16ul;	// Recieve up to 16KB blocks from the socket
+	static const size_t	ZLIB_CHUNK_SIZE	= 1024u;			// Chunk size for ZLib compression/decompression
 
 	// Static means you can call CBuffer::ReverseBuffer without having a CBuffer object at all
 	static void ReverseBuffer(const void* pInput, void* pOutput, size_t nLength);
