+++ shareaza/CtrlChatFrame.cpp
@@ -273,7 +273,7 @@ void CChatFrame::MoveHistory(int nDelta)
 	}
 
 	m_nHistory += nDelta;
-	m_nHistory = max( 0, min( m_pHistory.GetSize(), m_nHistory ) );
+	m_nHistory = max( 0, min( (int)m_pHistory.GetSize(), m_nHistory ) );
 
 	if ( m_nHistory == m_pHistory.GetSize() )
 	{
