+++ shareaza/PageSettingsConnection.cpp
@@ -302,7 +302,7 @@ void CConnectionSettingsPage::OnOK()
 	{
 		QWORD nDownload = max( Settings.Bandwidth.Downloads, Settings.Connection.InSpeed * Kilobits / Bytes );
 		QWORD nUpload   = Settings.Connection.OutSpeed * Kilobits / Bytes;
-		if ( Settings.Bandwidth.Uploads > 0 ) nUpload =  min( Settings.Bandwidth.Uploads, nUpload );
+		if ( Settings.Bandwidth.Uploads > 0 ) nUpload =  min( (QWORD)Settings.Bandwidth.Uploads, nUpload );
 
 		if ( nUpload * 16 < nDownload )
 		{
